Refactors Reducers Into Their Own File
